+++ startinsights/website_api/investor_wise_overview.py
@@ -81,51 +81,3 @@ def get_random_color_code():
     color = random.randrange(0, 2**24)
     hex_color = hex(color)
     return hex_color
-
-
-#after created the rounded wise form to show the list view
-@frappe.whitelist()
-def get_investor_wise_list_view(user_id):
-    try:
-        get_investor_wise = frappe.db.get_all("Investor-Wise Overview",{'creation_person_id':user_id},['*']) 
-        formatted_investor_wise= [] 
-        for investor in get_investor_wise:
-            if investor.share_certificate:
-                attach = get_url + investor.share_certificate
-            else:
-                attach = ""    
-            investor_wise = {
-                "investor_name":investor.investor_name,
-                "tag_name":investor.tag_name,
-                "date_of_allotment":investor.date_of_allotment,
-                "invested_round":investor.invested_round,
-                "amount_invested":investor.amount_invested,
-                "distinctive_share_no":investor.distinctive_share_no,
-                "share_certificate":attach,
-                "shares_allotted":investor.shares_allotted,
-                "price_per_share":investor.price_per_share,
-                "fully_diluted_shares":investor.fully_diluted_shares,
-                "class_of_shares":investor.class_of_shares,
-                "folio_number":investor.folio_number,
-                "shareholding":investor._shareholding,
-                "creation_person_id":investor.creation_person_id
-            }
-            formatted_investor_wise.append(investor_wise)
-        return {"status":True,"round_wise":formatted_investor_wise}
-    except Exception as e:
-        return {"status":False,"message":e} 
-
-
-#the below method to show the graph percentage and color code
-@frappe.whitelist()
-def get_investor_wise_graph(user_id):
-    try:
-        get_investor_wise_data = frappe.db.get_value("Investor-Wise Overview",{'creation_person_id':user_id},['investor_name','_shareholding','color_code'])
-        investor_wise_graph = {
-            "name":get_investor_wise_data[0],
-            "percentage":get_investor_wise_data[1],
-            "color_code":get_investor_wise_data[2]
-        }
-        return {"status":True,"investor_wise_graph":investor_wise_graph}
-    except Exception as e:
-        return {"status":False,"message":e}
