+++ startinsights/website_api/round_wise_overview.py
@@ -57,41 +57,3 @@ def get_random_color_code():
     color = random.randrange(0, 2**24)
     hex_color = hex(color)
     return hex_color
-
-#after created the rounded wise form to show the list view
-@frappe.whitelist()
-def get_round_wise_list_view(user_id):
-    try:
-        get_round_wise = frappe.db.get_all("Round-wise Overview",{'creation_person_id':user_id},['*']) 
-        formatted_round_wise= [] 
-        for round_wise in get_round_wise:
-            round_wise = {
-                "name_of_the_round":round_wise.name_of_the_round,
-                "round_type":round_wise.round_type,
-                "closing_date_of_the_round":round_wise.closing_date_of_the_round,
-                "description":round_wise.description,
-                "select_security_prefix":round_wise.select_security_prefix,
-                "amount_raised":round_wise.amount_raised,
-                "price_per_share":round_wise.price_per_share,
-                "pre_money_valuation":round_wise.pre_money_valuation,
-                "dilution_for_this_round":round_wise.dilution_for_this_round,
-                "creation_person_id":round_wise.creation_person_id
-            }
-            formatted_round_wise.append(round_wise)
-        return {"status":True,"round_wise":formatted_round_wise}
-    except Exception as e:
-        return {"status":False,"message":e} 
-
-#the below method to show the graph percentage and color code
-@frappe.whitelist()
-def get_round_wise_graph(user_id):
-    try:
-        get_round_wise_data = frappe.db.get_value("Round-wise Overview",{'creation_person_id':user_id},['name_of_the_round','dilution_for_this_round','color_code'])
-        round_wise_graph = {
-            "name":get_round_wise_data[0],
-            "percentage":get_round_wise_data[1],
-            "color_code":get_round_wise_data[2]
-        }
-        return {"status":True,"round_wise_graph":round_wise_graph}
-    except Exception as e:
-        return {"status":False,"message":e}
